修改加密密钥

+++ clever-system/src/main/java/com/clever/service/impl/EmailSubjectServiceImpl.java
@@ -2,7 +2,13 @@ package com.clever.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.clever.Constant;
 import com.clever.bean.model.OnlineUser;
+import com.clever.bean.system.EmailTemplate;
+import com.clever.exception.BaseException;
+import com.clever.exception.ConstantException;
+import com.clever.mapper.EmailTemplateMapper;
+import com.clever.util.DesUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -29,6 +35,9 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
 
     @Resource
     private EmailSubjectMapper emailSubjectMapper;
+
+    @Resource
+    private EmailTemplateMapper emailTemplateMapper;
 
     /**
      * 分页查询邮箱主体列表
@@ -89,13 +98,43 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void save(EmailSubject emailSubject, OnlineUser onlineUser) {
+        QueryWrapper<EmailSubject> queryWrapper = getEmailSubjectQueryWrapper(emailSubject, onlineUser);
+        if (StringUtils.isNotBlank(emailSubject.getId())) {
+            queryWrapper.ne("id", emailSubject.getId());
+        }
+        if (emailSubjectMapper.selectCount(queryWrapper) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("您已存在同一个邮箱主体"));
+        }
         if (StringUtils.isBlank(emailSubject.getId())) {
+            emailSubject.setPassword(DesUtil.safeEncrypt(emailSubject.getPassword(), Constant.KEY));
             emailSubjectMapper.insert(emailSubject);
             log.info("邮箱主体, 邮箱主体信息创建成功: userId={}, emailSubjectId={}", onlineUser.getId(), emailSubject.getId());
         } else {
+            EmailSubject oldEmailSubject = emailSubjectMapper.selectById(emailSubject.getId());
+            if (oldEmailSubject == null) {
+                throw new BaseException(ConstantException.DATA_NOT_EXIST.format("邮箱主体"));
+            }
+            // 防止重复加密
+            if (!oldEmailSubject.getPassword().equals(emailSubject.getPassword())) {
+                emailSubject.setPassword(DesUtil.safeEncrypt(emailSubject.getPassword(), Constant.KEY));
+            }
+
             emailSubjectMapper.updateById(emailSubject);
             log.info("邮箱主体, 邮箱主体信息修改成功: userId={}, emailSubjectId={}", onlineUser.getId(), emailSubject.getId());
         }
+    }
+
+    private static QueryWrapper<EmailSubject> getEmailSubjectQueryWrapper(EmailSubject emailSubject, OnlineUser onlineUser) {
+        QueryWrapper<EmailSubject> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("platform_id", emailSubject.getPlatformId());
+        queryWrapper.eq("creator", onlineUser.getId());
+        queryWrapper.eq("host", emailSubject.getHost());
+        queryWrapper.eq("driver", emailSubject.getDriver());
+        queryWrapper.eq("port", emailSubject.getPort());
+        queryWrapper.eq("encryption", emailSubject.getEncryption());
+        queryWrapper.eq("username", emailSubject.getUsername());
+        queryWrapper.eq("subject_name", emailSubject.getSubjectName());
+        return queryWrapper;
     }
 
     /**
@@ -106,31 +145,10 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void delete(String id, OnlineUser onlineUser) {
+        if (emailTemplateMapper.selectCount(new QueryWrapper<EmailTemplate>().eq("subject_id", id)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.format("该邮箱主体下还有邮件模板正在使用!!!"));
+        }
         emailSubjectMapper.deleteById(id);
         log.info("邮箱主体, 邮箱主体信息删除成功: userId={}, emailSubjectId={}", onlineUser.getId(), id);
-    }
-
-    /**
-     * 根据id列表删除邮箱主体信息
-     *
-     * @param ids        id列表
-     * @param onlineUser 当前登录用户
-     */
-    @Override
-    public void deleteBatchIds(List<String> ids, OnlineUser onlineUser) {
-        emailSubjectMapper.deleteBatchIds(ids);
-        log.info("邮箱主体, 邮箱主体信息批量删除成功: userId={}, count={}, emailSubjectIds={}", onlineUser.getId(), ids.size(), ids.toString());
-    }
-
-    /**
-     * 根据平台id删除邮箱主体
-     *
-     * @param platformId 平台id
-     * @param onlineUser 当前登录用户
-     */
-    @Override
-    public void deleteByPlatformId(String platformId, OnlineUser onlineUser) {
-        emailSubjectMapper.delete(new QueryWrapper<EmailSubject>().eq("platform_id", platformId));
-        log.info("邮箱主体, 邮箱主体信息根据平台id删除成功: userId={}, platformId={}", onlineUser.getId(), platformId);
     }
 }
